+++ backend/app/routes/transactions/router.py
@@ -24,14 +24,14 @@ from app.schemas.transaction import (
     TransactionsOverview,
     UpdateTransactionRequest,
 )
-from app.services.transactions.creation import create_transaction_and_get_response
-from app.services.transactions.deletion import delete_transaction_for_user
-from app.services.transactions.detail import get_transaction_response_for_user
-from app.services.transactions.imports import (
+from app.services.importers import (
     import_firefly_budgets,
     import_firefly_transactions,
     import_transactions,
 )
+from app.services.transactions.creation import create_transaction_and_get_response
+from app.services.transactions.deletion import delete_transaction_for_user
+from app.services.transactions.detail import get_transaction_response_for_user
 from app.services.transactions.listing import list_transaction_responses
 from app.services.transactions.overview import get_transactions_overview as get_transactions_overview_response
 from app.services.transactions.update import update_transaction_and_get_response
