+++ backend/app/services/importers/__init__.py
@@ -1,0 +1,12 @@
+"""Importers that turn uploaded exports into the user's records
+
+One package per import source, generic CSV and Firefly III, with the
+machinery they both build on in shared. Each source package owns only what
+is specific to its export format
+"""
+
+from app.services.importers.firefly.budgets import import_firefly_budgets
+from app.services.importers.firefly.service import import_firefly_transactions
+from app.services.importers.generic.service import import_transactions
+
+__all__ = ["import_firefly_budgets", "import_firefly_transactions", "import_transactions"]
