+++ backend/app/services/importers/shared/categories.py
@@ -10,8 +10,8 @@ from app.models.category import Category
 from app.models.group import GroupMember
 from app.models.user import User
 from app.schemas.transaction import TransactionImportCategoryMapping, TransactionImportCreateCategory
-from app.services.transactions.imports.shared.stats import ImportStats
-from app.services.transactions.imports.shared.validation_helpers import strip_import_text_or_raise
+from app.services.importers.shared.stats import ImportStats
+from app.services.importers.shared.validation_helpers import strip_import_text_or_raise
 
 
 async def get_or_create_import_categories_by_source(
