--- backend/app/services/transactions/imports/__init__.py
@@ -1,12 +1,0 @@
-"""Transaction import services
-
-One package per import source, generic CSV and Firefly III, with the
-machinery they both build on in shared. Each source package owns only what
-is specific to its export format
-"""
-
-from app.services.transactions.imports.firefly.budgets import import_firefly_budgets
-from app.services.transactions.imports.firefly.service import import_firefly_transactions
-from app.services.transactions.imports.generic.service import import_transactions
-
-__all__ = ["import_firefly_budgets", "import_firefly_transactions", "import_transactions"]
